pyproject.toml
[project]
name = "snoop-edwards"
version = "0.1.0"
description = "Project managing a main app, a Streamlit app, and a Discord bot"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Guillermo Rodriguez", email = "guillermo.rodriguez@snoopconsulting.com" }
]


dependencies = [
    "streamlit>=1.45.1",
    "requests>=2.32.3",
    "discord>=2.3.2",
    "quart>=0.20.0",
    "quart-cors>=0.8.0",
    "openai>=1.84.0",
    "httpx>=0.28.1",
    "tiktoken>=0.9.0",
    "twilio>=9.6.2",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "pandas>=2.2.3",
    "azure-cognitiveservices-speech>=1.44.0",
    "pydub>=0.25.1",
    "azure-storage-blob>=12.25.1",
    "azure-search-documents>=11.5.2",
    "azure-identity>=1.23.0",
    "google-cloud-bigquery>=3.34.0",
    "langsmith>=0.3.44",
    "langgraph",
    "langgraph-cli[inmem]>=0.3.4",
    "langchain-openai>=0.3.19",
    "langchain-community>=0.3.24",
    "faiss-cpu>=1.11.0",
    "google-cloud-bigquery-storage>=2.32.0",
    "sqlalchemy>=2.0.41",
    "sqlalchemy-bigquery",
    "pydantic[email]>=2.11.5",
    "nltk>=3.9.1",
    "psycopg2",
    "google-cloud-aiplatform>=1.95.1",
    "langgraph-checkpoint-postgres>=2.0.21",
    "langchain-pinecone>=0.2.6",
    "langchain-docling>=0.2.0",
    "langchain-mcp-adapters>=0.1.4",
    "google-api-python-client>=2.170.0",
    "langchain-google-genai>=2.1.5",
    "langmem>=0.0.28",
    "python-dotenv>=1.1.1",
    "langchain-google-vertexai>=2.0.28",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
